==> Address clustering - ML/1) Preprocess data.ipynb <==
import pandas as pd
import numpy as np

import collections
import random

import gc

import utils

date_string = "20210720"

### Import predictions

df_predictions = pd.read_csv(utils.get_base_path("heuristics/{}_smart_prediction.csv".format(date_string)), index_col=0)

df_predictions.columns

# drop pow10, because we can use the individual pow10 heuristics with random forest
# drop two_hops, because it may be biased towards our ground truth set (due to higher rate of clustering)
df_predictions.drop(columns=["pow10", "two_hops"], inplace=True)

### Create a boolean mask for outputs where the transaction has at least one vote

has_prediction_full = (df_predictions.drop(columns=["true_change"]) != -1).any(1)

has_prediction_full_outputs = np.array([x for x in has_prediction_full for _ in (0,1)])

del has_prediction_full

FP_COLUMNS = [x for x in df_predictions.columns if x[:3] == "fp_"]

FP_COLUMNS

has_prediction_nofp = (df_predictions.drop(columns=(["true_change"] + FP_COLUMNS)) != -1).any(1)

has_prediction_nofp_outputs = np.array([x for x in has_prediction_nofp for _ in (0,1)])

del has_prediction_nofp

np.save(utils.get_base_path("evaluation/{0}_mask.npy".format(date_string)), has_prediction_full_outputs)

np.save(utils.get_base_path("evaluation/{0}_mask_nofp.npy".format(date_string)), has_prediction_nofp_outputs)

del has_prediction_full_outputs, has_prediction_nofp_outputs

### Convert tx-based predictions into output-based predictions

Input: index of output that heuristic votes for, or -1 otherwise.

Output: True/False (as 1/0) if the respective output is the change, or -1 otherwise

for row in df_predictions.itertuples():
    assert list(row._fields)[1:] == list(df_predictions.columns)
    break

values, indexes = [], []

for row in df_predictions.itertuples():
    indexes.append(row[0]*10)
    indexes.append(row[0]*10+1)
    values.append([int(x == 0) if x != -1 else -1 for x in row[1:]])
    values.append([int(x == 1) if x != -1 else -1 for x in row[1:]])

df_predictions_outs = pd.DataFrame(values, index=indexes)

df_predictions_outs.columns = df_predictions.columns

df_predictions_outs.head(6)

df_predictions.head(3)

del df_predictions

del values, indexes

gc.collect()

### Split into features and outcome

y = df_predictions_outs["true_change"]

y = y.astype("uint8")

np.save(utils.get_base_path("evaluation/{0}_outcomes.npy".format(date_string)), y.values)

X = df_predictions_outs.drop("true_change", axis=1)

# reorder variables
X = X.replace([0, -1], [-1, 0])

X = X.apply(lambda x: x.astype('category'))

X.head()

del df_predictions_outs

### Concatenate with tx/output characteristics

See notebook 9a) for export of characteristics

df_output_dtypes = {
    "co_output_value": np.uint64,
    "co_output_value_ratio": np.float64,
    "co_is_larger_output": np.bool,
    "co_output_index": np.uint8,
    "ct_fee": np.uint64,
    "ct_fee_per_byte": np.uint64,
    "ct_tx_value": np.uint64,
    "ct_segwit_tx": np.uint8,
    "ct_has_locktime": np.bool,
    "ct_block_height": np.uint32,
    "ct_input_count": np.uint32,
    "co_fresh_output": np.bool,
    "co_other_fresh": np.bool
}

df_outputs = pd.read_csv(utils.get_base_path("heuristics/{}-output-features.csv".format(date_string)), index_col=0, dtype=df_output_dtypes)

assert len(X) == len(df_outputs)

df_outputs.index = X.index

X = pd.concat([X, df_outputs], axis=1)

del df_outputs

len(X)

### Export

X.to_csv(utils.get_base_path("heuristics/{}-combined.csv".format(date_string)))
