==> Address clustering - ML/3d) Persist RF Model.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns

from sklearn.ensemble import RandomForestClassifier

import joblib

import utils

sns.set_theme(style="ticks")

import random

import gc

date_string = "20210720"

### Import data

y = np.load(utils.get_base_path("evaluation/{}_outcomes.npy".format(date_string)))

X = pd.read_csv(utils.get_base_path("evaluation/{}-combined.csv".format(date_string)), index_col=0, dtype=utils.get_X_dtypes())

X["ct_epoch"] = X["ct_block_height"] // 1008

X.columns

X.drop(columns=["ct_fee", "co_output_value", "ct_block_height", "co_is_larger_output", "co_fresh_output", "co_other_fresh"], inplace=True)

X_cols = X.columns

len(X.columns)

X_cols

FINGERPRINT_COLS = [x for x in X_cols if x[:3] == "fp_"]

CHARACTERISTICS_COLS = [x for x in X_cols if x[:3] == "co_" or x[:3] == "ct_"]

### Remove transactions that do not have a single prediction

mask_full = (X.drop(columns=CHARACTERISTICS_COLS) != 0).any(1)

assert (mask_full == np.load(utils.get_base_path("evaluation/{}_mask.npy".format(date_string)))).all()

np.sum(mask_full) // 2

len(X) // 2

assert len(mask_full) == len(X)

X1 = X[mask_full].copy()
y1 = y[mask_full]

assert len(X1) == np.sum(mask_full)

### Train classifier for normal & fingerprint

rfc1 = RandomForestClassifier(random_state=1337, n_jobs=34, n_estimators=100, max_features=7, min_samples_split=20, min_samples_leaf=10)

rfc1.fit(X1, y1)

joblib.dump(X1.columns, utils.get_base_path('model/X_columns_full.pkl'))

joblib.dump(rfc1, utils.get_base_path("model/rfc-full.pkl"))

del X1, y1

del rfc1

### Select only transactions that have a prediction from a "normal" heuristic

X2 = X.copy()
y2 = y

X2 = X2.drop(columns=FINGERPRINT_COLS)

len(X2)

mask_nofp = (X2.drop(columns=CHARACTERISTICS_COLS) != 0).any(1)

assert (mask_nofp == np.load(utils.get_base_path("evaluation/{}_mask_nofp.npy".format(date_string)))).all()

np.sum(mask_nofp) // 2

assert len(mask_nofp) == len(X2)
assert len(mask_nofp) == len(y2)

X2 = X2[mask_nofp]
y2 = y2[mask_nofp]

assert len(X2) == np.sum(mask_nofp)
assert len(y2) == np.sum(mask_nofp)

### Train classifier for normal only

rfc2 = RandomForestClassifier(random_state=1337, n_jobs=34, n_estimators=100, max_features=6, min_samples_split=20, min_samples_leaf=10)

rfc2.fit(X2, y2)

joblib.dump(X2.columns, utils.get_base_path('model/X_columns_nofp.pkl'))

joblib.dump(rfc2, utils.get_base_path("model/rfc-nofp.pkl"))

del X2, y2

del rfc2
