==> Address clustering - ML/4c) Repeated RF ROC - Universal Only.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns

from sklearn.ensemble import RandomForestClassifier

from sklearn.model_selection import KFold, GroupKFold, cross_val_score, train_test_split, ShuffleSplit, GroupShuffleSplit

from sklearn.metrics import roc_curve, plot_roc_curve, RocCurveDisplay, auc, roc_auc_score

from sklearn.experimental import enable_halving_search_cv

from sklearn.model_selection import HalvingGridSearchCV

sns.set_theme(style="ticks")

import random

import collections

import gc

import utils

date_string = "20210720"

### Import data

y = np.load(utils.get_base_path("evaluation/{}_outcomes.npy".format(date_string)))

X = pd.read_csv(utils.get_base_path("evaluation/{}-combined.csv".format(date_string)), index_col=0, dtype=utils.get_X_dtypes())

X["ct_epoch"] = X["ct_block_height"] // 1008

X.columns

X.drop(columns=["ct_fee", "co_output_value", "ct_block_height", "co_is_larger_output", "co_fresh_output", "co_other_fresh"], inplace=True)

X_cols = X.columns

FINGERPRINT_COLS = [x for x in X_cols if x[:3] == "fp_"]

X.drop(columns=FINGERPRINT_COLS, inplace=True)

mask_nofp = np.load(utils.get_base_path("evaluation/{}_mask_nofp.npy".format(date_string)))

np.sum(mask_nofp) // 2

len(X) // 2

assert len(mask_nofp) == len(X)

X = X[mask_nofp].copy()
y = y[mask_nofp]

assert len(X) == np.sum(mask_nofp)
assert len(y) == np.sum(mask_nofp)

cluster_ids = np.load(utils.get_base_path("heuristics/{}-cluster-ids.npy".format(date_string)))

assert len(cluster_ids) == len(mask_nofp)

cluster_groups = cluster_ids[mask_nofp]

assert len(cluster_groups) == np.sum(mask_nofp)

### Apply to Out of Bag Sample (without having to rerun previous code)

shuffler = GroupShuffleSplit(n_splits=21, test_size=.2, random_state=1337).split(X=X, groups=cluster_groups)

true_outcomes = []
predictions = []

iteration = 0
for ib_idx, oob_idx in shuffler:
    assert len(ib_idx) + len(oob_idx) == len(X)
    assert len(ib_idx) > len(oob_idx)
    
    X_ib = X.iloc[ib_idx].copy()
    y_ib = y[ib_idx]
    X_oob  = X.iloc[oob_idx].copy()
    y_oob  = y[oob_idx]
    
    clusters_ib = cluster_groups[ib_idx]
    clusters_oob = cluster_groups[oob_idx]
    
    assert len(X_ib) == len(y_ib)
    assert len(X_oob) == len(y_oob)
    assert len(X_ib) + len(X_oob) == len(X)
    assert len(set(X_ib.index).intersection(set(X_oob.index))) == 0
    assert len(set(clusters_ib).intersection(set(clusters_oob))) == 0
    
    rfc = RandomForestClassifier(random_state=1337+iteration, n_jobs=34, n_estimators=100, max_features=6, min_samples_split=20, min_samples_leaf=10)
    rfc.fit(X_ib, y_ib)
    oob_prediction = rfc.predict_proba(X_oob)[:, 1]
    true_outcomes.append(y_oob)
    predictions.append(oob_prediction)
    print(iteration, roc_auc_score(y_oob, oob_prediction))
    iteration += 1

len(predictions)

auc_scores = [roc_auc_score(y_true, y_pred) for y_true, y_pred in zip(true_outcomes, predictions)]

auc_scores

np.mean(auc_scores[1:]), np.std(auc_scores[1:])
